requirements.txt
Flask==2.3.2
requests==2.31.0
Werkzeug==2.3.6
flask-cors==3.0.10
msal==1.22.0
python-dotenv==1.0.1